nagoyameshi-django-api/requirements.txt
asgiref
Django
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
pytz
sqlparse